src/requirements.txt
absl-py==0.10.0
affine==2.4.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.5.2
asttokens==3.0.0
astunparse==1.6.3
async-timeout==3.0.1
backports.functools-lru-cache==1.6.1
blinker==1.8.2
boto==2.49.0
bz2file==0.98
certifi==2020.6.20
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.3
contextily==1.5.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
et-xmlfile==1.0.1
exceptiongroup==1.2.2
executing==2.2.0
fake-useragent==2.0.0
fastapi==0.115.6
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.16.1
fiona==1.10.1
flask==3.0.3
fsspec==2024.12.0
gast==0.3.3
gdown==5.2.0
geographiclib==2.0
geopandas==0.13.2
geopy==2.4.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gradio==4.44.1
gradio-client==1.3.0
gunicorn==23.0.0
h11==0.14.0
h5py==2.10.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipython==8.12.3
itsdangerous==2.2.0
jdcal==1.4.1
jinja2==3.1.5
joblib==1.4.2
jupyter-contrib-core==0.3.3
jupyter-core==5.7.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
lazy-loader==0.4
lime==0.2.0.1
LunarCalendar==0.0.9
Markdown==3.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mercantile==1.2.1
mistune==3.1.0
mkl-fft==1.2.0
mkl-random==1.1.0
mkl-service==2.3.0
mpmath==1.3.0
nbconvert==7.16.5
nbformat==5.10.4
networkx==3.1
nltk==3.9.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.11.0.86
opt-einsum==3.3.0
orc==0.1.0
orjson==3.10.15
osqp==0.6.1
packaging==24.2
pandocfilters==1.4.2
parso==0.7.0
patsy==0.5.1
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt-toolkit==3.0.50
protobuf==3.11.2
ptyprocess==0.6.0
pure-eval==0.2.3
pyarrow==0.15.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pybind11==2.5.0
pydantic==2.10.5
pydantic-core==2.27.2
pydub==0.25.1
pygments==2.19.1
pyogrio==0.9.0
pyOpenSSL==19.1.0
pyproj==3.5.0
pystan==2.19.1.1
python-dateutil==2.8.1
python-multipart==0.0.20
python-version==0.0.2
pytz==2020.1
PyWavelets==1.4.1
pywin32==308
PyYAML==5.3.1
pyzmq==19.0.2
QtPy==1.9.0
rasterio==1.3.11
regex==2024.11.6
requests-oauthlib==1.3.0
rich==13.9.4
ruff==0.9.2
scikeras==0.12.0
scikit-image==0.21.0
scikit-learn==1.3.2
scikit-surprise==1.1.1
semantic-version==2.10.0
Send2Trash==1.5.0
setuptools-git==1.2
shapely==2.0.7
shellingham==1.5.4
simplegeneric==0.8.1
sip==4.19.13
sniffio==1.3.1
snuggs==1.4.7
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.3
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
testpath==0.4.4
textblob==0.18.0.post0
tifffile==2023.7.10
tomlkit==0.12.0
torch==2.4.1
torchvision==0.19.1
typer==0.15.1
typing-extensions==4.12.2
urllib3==1.25.11
uvicorn==0.33.0
vaderSentiment==3.3.2
websockets==12.0
werkzeug==3.0.6
wincertstore==0.2
wordcloud==1.9.4
wrapt==1.12.1
xgboost==1.2.1
xmltodict==0.14.2
xyzservices==2025.1.0
yarg==0.1.9
zipp==3.20.2
